Use reflection for Require(), Include()
When this was originally written reflection came with rather a lot of
downsides, such as disabling dead code elimination. Since then, things
have improved.

For Required() I want to support:

	type MyType string

	v.Required("key", MyType("..."))

And also just simplify the code.

Similar for Include():

	type MyType string

	v.Include("key", MyType("..."), []MyType{...})

The "any" here is rather annoying, but can't have type parameters on
methods and can't really think of a better way to do it.